SUMA: Move clipping plane functions to separate .c/.h files
